# specs/003-wallet-import/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WalletSeed Statement Extraction API
  version: 0.3.0
  description: |
    HTTP API for the WalletSeed application. This contract is the
    Wallet-Import evolution of `002-multi-bank-support`'s OpenAPI surface.
    Every change versus 0.2.0 is **additive**:

      * Four new paths under `/api/wallet/...` are introduced:
        - `GET  /api/wallet/accounts`
        - `GET  /api/wallet/categories`
        - `POST /api/wallet/import/compare`
        - `POST /api/wallet/import/submit`
      * One new error envelope (`WalletErrorResponse`) and one new
        error-code enum (`WalletErrorCode`) are introduced — used **only**
        by the new paths above. The existing `ExtractionErrorResponse`
        and `ErrorCodes` enum are unchanged and continue to be the
        contract for the existing `/api/statements/extract` path AND for
        the PDF-stage failures of `POST /api/wallet/import/compare`
        (size, magic-byte, empty, unrecognized layout, etc.).
      * The existing path `POST /api/statements/extract` and **every**
        schema introduced by 0.1.0 / 0.2.0 (ExtractedStatementResponse,
        StatementHeader, CardholderSection, Transaction, totals,
        ReconciliationStatus, ExtractionErrorResponse, BankInfo, etc.)
        are byte-identical to 0.2.0.

    The Wallet integration is server-side only: the backend proxies all
    calls to the upstream Wallet API; the frontend never holds the
    Wallet JWT.

servers:
  - url: http://localhost:5080
    description: Local development

paths:

  # ============================================================
  # Unchanged from 0.2.0
  # ============================================================
  /api/statements/extract:
    post:
      summary: Extract transactions from a credit-card statement PDF (auto-detected bank)
      description: |
        Unchanged from 0.2.0. See the 0.2.0 contract for the full
        narrative; the behaviour and response shape are preserved
        byte-for-byte. Reproduced here only so this document stands
        alone.
      operationId: extractStatement
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Statement parsed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractedStatementResponse'
        '400':
          description: INVALID_FILE_TYPE or EMPTY_FILE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionErrorResponse'
        '413':
          description: FILE_TOO_LARGE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionErrorResponse'
        '422':
          description: PASSWORD_PROTECTED, NO_TEXT_EXTRACTABLE, UNRECOGNIZED_LAYOUT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionErrorResponse'
        '500':
          description: PARSE_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionErrorResponse'

  # ============================================================
  # New in 0.3.0
  # ============================================================
  /api/wallet/accounts:
    get:
      summary: List Wallet accounts (proxy)
      description: |
        Fetches Wallet accounts via the backend's configured Wallet
        credentials. Archived accounts are filtered out before returning.

        The backend reads the Wallet JWT from local configuration
        (`appsettings:Wallet:Jwt` or env `Wallet__Jwt`). The frontend
        never sees the JWT.
      operationId: listWalletAccounts
      responses:
        '200':
          description: Accounts returned.
          content:
            application/json:
              schema:
                type: object
                required: [accounts]
                properties:
                  accounts:
                    type: array
                    items: { $ref: '#/components/schemas/WalletAccount' }
        '502':
          description: WALLET_CREDENTIALS_INVALID (Wallet API returned 401/403) or WALLET_REJECTED (Wallet API returned 4xx other than 401/403)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletErrorResponse' }
        '503':
          description: WALLET_NOT_CONFIGURED — backend has no Wallet JWT configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletErrorResponse' }
        '504':
          description: WALLET_UNAVAILABLE — Wallet API timeout, network failure, or 5xx
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletErrorResponse' }

  /api/wallet/categories:
    get:
      summary: List Wallet categories (proxy)
      operationId: listWalletCategories
      responses:
        '200':
          description: Categories returned.
          content:
            application/json:
              schema:
                type: object
                required: [categories]
                properties:
                  categories:
                    type: array
                    items: { $ref: '#/components/schemas/WalletCategory' }
        '502': { description: WALLET_CREDENTIALS_INVALID or WALLET_REJECTED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '503': { description: WALLET_NOT_CONFIGURED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '504': { description: WALLET_UNAVAILABLE, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }

  /api/wallet/import/compare:
    post:
      summary: Extract PDF, fetch Wallet records in the period, and return a side-by-side comparison
      description: |
        Accepts a single PDF and an `accountId` and returns the
        comparison view's source-of-truth payload:

          - The PDF's transactions (one row per extracted Transaction),
            with each row's matched Wallet record ids (per the
            duplicate-detection rule: ±2 calendar days + exact amount +
            same currency + same direction).
          - The chosen Wallet account's existing records that fall
            within the PDF's statement period extended by ±5 days at
            each boundary (fully paginated).
          - The list of Wallet categories (so the frontend can show a
            per-row category dropdown without a second round trip).
          - The chosen Wallet account's metadata.
          - The list of cardholder section names that have no entry in
            the backend's label-mapping configuration.

        The PDF-stage error envelope is the existing
        `ExtractionErrorResponse` (FR-006); the Wallet-stage error
        envelope is `WalletErrorResponse` (new).
      operationId: walletImportCompare
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, accountId]
              properties:
                file:
                  type: string
                  format: binary
                accountId:
                  type: string
                  description: Id of the chosen Wallet account.
      responses:
        '200':
          description: Comparison ready.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompareResponse' }
        '400':
          description: INVALID_FILE_TYPE or EMPTY_FILE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExtractionErrorResponse' }
        '413':
          description: FILE_TOO_LARGE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExtractionErrorResponse' }
        '422':
          description: PASSWORD_PROTECTED, NO_TEXT_EXTRACTABLE, UNRECOGNIZED_LAYOUT
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExtractionErrorResponse' }
        '502': { description: WALLET_CREDENTIALS_INVALID or WALLET_REJECTED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '503': { description: WALLET_NOT_CONFIGURED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '504': { description: WALLET_UNAVAILABLE, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '500':
          description: PARSE_FAILED (extraction crash). Wallet-side crashes go through 502/504 instead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExtractionErrorResponse' }

  /api/wallet/import/submit:
    post:
      summary: Create the selected records in Wallet, in batches of up to 50
      description: |
        Each row in `rows` becomes one Wallet record. The backend:

          1. Resolves `labelIds` for each row from the configured
             cardholder-section → label-id mapping.
          2. Sets `paymentType = "credit_card"` for every row.
          3. Chunks rows into groups of ≤50, in input order.
          4. POSTs each chunk to Wallet `POST /v1/api/records`.
          5. Reports per-row outcomes.

        A row failure does not block other rows (FR-019). A 401/403 on
        any chunk stops further chunks and surfaces as a single
        `WALLET_CREDENTIALS_INVALID` envelope (FR-020).
      operationId: walletImportSubmit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubmitRequest' }
      responses:
        '200':
          description: Submission completed (possibly with per-row failures).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmitResponse' }
        '400':
          description: Request shape invalid (empty rows, missing categoryId on a row, missing accountId).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WalletErrorResponse' }
        '502': { description: WALLET_CREDENTIALS_INVALID or WALLET_REJECTED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '503': { description: WALLET_NOT_CONFIGURED, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }
        '504': { description: WALLET_UNAVAILABLE, content: { application/json: { schema: { $ref: '#/components/schemas/WalletErrorResponse' } } } }

components:
  schemas:

    # ============================================================
    # New in 0.3.0
    # ============================================================
    WalletAccount:
      type: object
      required: [id, name, currencyCode, accountType]
      properties:
        id:           { type: string }
        name:         { type: string }
        currencyCode: { type: string, example: "USD" }
        accountType:  { type: string, example: "CreditCard" }

    WalletCategory:
      type: object
      required: [id, name]
      properties:
        id:    { type: string }
        name:  { type: string }
        color: { type: [string, "null"] }

    StatementWindow:
      type: object
      required: [from, to, issueDate, cutoffDate]
      properties:
        from:       { type: string, format: date, description: "Inclusive lower bound of the Wallet records fetch window (issueDate - 5d)." }
        to:         { type: string, format: date, description: "Inclusive upper bound of the Wallet records fetch window (cutoffDate + 5d)." }
        issueDate:  { type: string, format: date }
        cutoffDate: { type: string, format: date }

    PdfRow:
      type: object
      required:
        - index
        - date
        - signedAmount
        - currency
        - description
        - cardholderSectionRawName
        - cardLast4
        - matchedWalletRecordIds
        - defaultSelected
        - currencyMismatch
      properties:
        index:
          type: integer
          minimum: 0
          description: 0-based row index in document order. Stable across Compare/Submit.
        date:
          type: string
          format: date
        signedAmount:
          type: number
          description: Negative for expense, positive for income.
        currency:
          type: string
          description: Currency code. In this iteration always equals the chosen Wallet account's currency.
          example: "USD"
        description:
          type: string
        counterParty:
          type: [string, "null"]
        cardholderSectionRawName:
          type: string
          description: The PDF cardholder section's RawName — used by the backend to resolve labels.
        cardLast4:
          type: string
          pattern: '^[0-9]{4}$'
        matchedWalletRecordIds:
          type: array
          items: { type: string }
          description: |
            Wallet record ids that this PDF row matches by the duplicate
            rule (±2 days + exact amount + same currency + same direction).
            Sorted lexicographically. Empty array means no duplicate.
        defaultSelected:
          type: boolean
          description: |
            Server-computed initial selection: true iff
            `matchedWalletRecordIds.length == 0 && !currencyMismatch`.
        currencyMismatch:
          type: boolean
          description: |
            True iff this row's currency differs from the chosen Wallet
            account's currency. Always false in this iteration; reserved
            for a future spec that adds per-row currency extraction.

    WalletRow:
      type: object
      required: [id, date, signedAmount, currency, claimedByPdfIndices]
      properties:
        id:           { type: string }
        date:         { type: string, format: date }
        signedAmount: { type: number }
        currency:     { type: string }
        note:         { type: [string, "null"] }
        counterParty: { type: [string, "null"] }
        categoryName: { type: [string, "null"] }
        claimedByPdfIndices:
          type: array
          items: { type: integer, minimum: 0 }
          description: Indices of PDF rows that matched this Wallet row. Sorted ascending.

    CompareResponse:
      type: object
      required:
        - window
        - account
        - categories
        - pdfRows
        - walletRows
        - unmappedSections
      properties:
        window:     { $ref: '#/components/schemas/StatementWindow' }
        account:    { $ref: '#/components/schemas/WalletAccount' }
        categories:
          type: array
          items: { $ref: '#/components/schemas/WalletCategory' }
        pdfRows:
          type: array
          items: { $ref: '#/components/schemas/PdfRow' }
          description: In document order (index ascending).
        walletRows:
          type: array
          items: { $ref: '#/components/schemas/WalletRow' }
          description: Sorted by (date ASC, id ASC).
        unmappedSections:
          type: array
          items: { type: string }
          description: |
            Distinct cardholder section RawNames whose label mapping is
            absent in the backend config. Sorted alphabetically.

    SubmitRequest:
      type: object
      required: [accountId, rows]
      properties:
        accountId: { type: string }
        rows:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/SubmitRequestRow' }

    SubmitRequestRow:
      type: object
      required: [index, date, signedAmount, currency, description, cardholderSectionRawName, categoryId]
      properties:
        index:                     { type: integer, minimum: 0 }
        date:                      { type: string, format: date }
        signedAmount:              { type: number }
        currency:                  { type: string }
        description:               { type: string }
        counterParty:              { type: [string, "null"] }
        cardholderSectionRawName:  { type: string }
        categoryId:                { type: string }

    SubmitResponse:
      type: object
      required: [outcomes]
      properties:
        outcomes:
          type: array
          items: { $ref: '#/components/schemas/SubmitOutcome' }

    SubmitOutcome:
      type: object
      required: [index, ok]
      properties:
        index:           { type: integer, minimum: 0 }
        ok:              { type: boolean }
        walletRecordId:  { type: [string, "null"], description: "Set when ok=true." }
        errorMessage:    { type: [string, "null"], description: "Set when ok=false; human-readable, safe to display." }

    WalletErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { $ref: '#/components/schemas/WalletErrorCode' }
            message: { type: string }

    WalletErrorCode:
      type: string
      enum:
        - WALLET_NOT_CONFIGURED
        - WALLET_CREDENTIALS_INVALID
        - WALLET_UNAVAILABLE
        - WALLET_REJECTED

    # ============================================================
    # Unchanged from 0.2.0
    # ============================================================
    ExtractedStatementResponse:
      type: object
      required:
        - statement
        - sections
        - totals
        - reconciliationStatus
        - needsReviewCount
        - unmappedCards
        - bank
      properties:
        statement:           { $ref: '#/components/schemas/StatementHeader' }
        sections:
          type: array
          items: { $ref: '#/components/schemas/CardholderSection' }
        totals:              { $ref: '#/components/schemas/StatementTotals' }
        reconciliationStatus: { $ref: '#/components/schemas/ReconciliationStatus' }
        needsReviewCount:    { type: integer, minimum: 0 }
        unmappedCards:
          type: array
          items: { type: string, pattern: '^[0-9]{4}$' }
        bank:                { $ref: '#/components/schemas/BankInfo' }

    BankInfo:
      type: object
      required: [id, displayName]
      properties:
        id:          { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,31}$' }
        displayName: { type: string }

    StatementHeader:
      type: object
      required: [cardType, maskedAccount, period, pageCount]
      properties:
        cardType:       { type: string }
        maskedAccount:  { type: string }
        period:         { $ref: '#/components/schemas/StatementPeriod' }
        pageCount:      { type: integer, minimum: 1 }

    StatementPeriod:
      type: object
      required: [issueDate, cutoffDate]
      properties:
        issueDate:  { type: string, format: date }
        cutoffDate: { type: string, format: date }

    CardholderSection:
      type: object
      required: [cardLast4, rawName, transactions, totals, reconciliationStatus]
      properties:
        cardLast4:   { type: string, pattern: '^[0-9]{4}$' }
        rawName:     { type: string }
        transactions:
          type: array
          items: { $ref: '#/components/schemas/Transaction' }
        totals:               { $ref: '#/components/schemas/SectionTotals' }
        reconciliationStatus: { $ref: '#/components/schemas/ReconciliationStatus' }

    Transaction:
      type: object
      required:
        - date
        - postingDate
        - referenceNumber
        - sequenceCode
        - rowType
        - description
        - amount
        - direction
        - cardLast4
        - needsReview
        - categoryId
        - categoryName
        - labelId
        - labelName
        - labelUnmapped
      properties:
        date:            { type: string, format: date }
        postingDate:     { type: string, format: date }
        referenceNumber: { type: string }
        sequenceCode:    { type: string, pattern: '^[CXP][0-9]+$' }
        rowType:         { type: string, enum: [purchase, financing, payment, adjustment] }
        description:     { type: string }
        amount:          { type: number, minimum: 0 }
        direction:       { type: string, enum: [income, expense] }
        cardLast4:       { type: string, pattern: '^[0-9]{4}$' }
        needsReview:     { type: boolean }
        categoryId:      { type: [string, "null"], format: uuid }
        categoryName:    { type: [string, "null"] }
        labelId:         { type: [string, "null"], format: uuid }
        labelName:       { type: [string, "null"] }
        labelUnmapped:   { type: boolean }

    SectionTotals:
      type: object
      required: [computedCharges, computedCredits, printedCharges, printedCredits]
      properties:
        computedCharges:  { type: number, minimum: 0 }
        computedCredits:  { type: number, minimum: 0 }
        printedCharges:   { type: [number, "null"] }
        printedCredits:   { type: [number, "null"] }

    StatementTotals:
      type: object
      required: [computedExpense, computedIncome, printedExpense, printedIncome]
      properties:
        computedExpense:  { type: number, minimum: 0 }
        computedIncome:   { type: number, minimum: 0 }
        printedExpense:   { type: [number, "null"] }
        printedIncome:    { type: [number, "null"] }

    ReconciliationStatus:
      type: string
      enum: [ok, needsReview]

    ExtractionErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_FILE_TYPE
                - EMPTY_FILE
                - FILE_TOO_LARGE
                - PASSWORD_PROTECTED
                - NO_TEXT_EXTRACTABLE
                - UNRECOGNIZED_LAYOUT
                - PARSE_FAILED
            message: { type: string }